116-FEAT: new image upload modal
