+++ apps/website/package.json
@@ -68,8 +68,6 @@
     "next": "^14.2.3",
     "next-axiom": "^1.2.0",
     "next-safe-action": "^6.2.0",
-    "next-seo": "^6.5.0",
-    "next-themes": "^0.3.0",
     "react": "^18.3.1",
     "react-beautiful-dnd": "^13.1.1",
     "react-dom": "^18.3.1",
